+++ AssignmentTwo/Model/Shapes/Circle.cs
@@ -11,7 +11,7 @@
         /// <param name="color">color</param>
         /// <param name="radius">radius</param>
         /// <param name="name">shape</param>
-        public Circle(string name, string color, double radius)
+        public Circle(string name, string color, decimal radius)
             : base(name, color)
         {
             this.Radius = radius;
@@ -21,17 +21,17 @@
         /// Gets or sets Radius
         /// </summary>
         /// <value>
-        /// A <see cref="double"/> containing the radius of the circle.
+        /// A <see cref="decimal"/> containing the radius of the circle.
         /// </value>
-        public double Radius { get; set; }
+        public decimal Radius { get; set; }
 
         /// <summary>
         /// THis is for calculating area
         /// </summary>
-        /// <returns>double</returns>
-        public override double GetArea()
+        /// <returns>decimal</returns>
+        public override decimal GetArea()
         {
-            return Math.PI * this.Radius * this.Radius;
+            return (decimal)Math.PI * this.Radius * this.Radius;
         }
     }
 }
